[Chore/#9] CheerType Model 분리
