+++ TheGreatestYourManito-iOS/TheGreatestYourManito-iOS/Presentation/Participate/PlayManitto/SubView/PlayManittoBottomView.swift
@@ -48,56 +48,6 @@ extension PlayManittoBottomView {
 }
 
 extension PlayManittoBottomView {
-    
-    enum CheerType: CaseIterable {
-        case luck
-        case love
-        case fire
-        case gift
-        
-        var defaultImage: ImageResource {
-            switch self {
-                
-            case .luck:
-                return .clover
-            case .love:
-                return .heart
-            case .fire:
-                return .fire
-            case .gift:
-                return .gift
-            }
-        }
-        
-        var clickedImage: ImageResource {
-            switch self {
-                
-            case .luck:
-                return .cloverClicked
-            case .love:
-                return .heartClicked
-            case .fire:
-                return .fireClicked
-            case .gift:
-                return .giftClicked
-            }
-        }
-        
-        var name: String {
-            switch self {
-                
-            case .luck:
-                return "luck"
-            case .love:
-                return "love"
-            case .fire:
-                return "fire"
-            case .gift:
-                return "present"
-            }
-        }
-    }
-    
     struct buttonStack: View {
         
         @Binding var cheerType: CheerType
